Dump debugging output for tests with "TBD" assertion
This makes it possible for test authors to see what sorts of API might
be appropriate to include in their assertions. In the case of property
assertions, they will have immediate access to the property values. In
the case of interface methods, they'll currently have the method name
and expected arguments. But also dumping result of sample call is coming
soon.