Reduce dependency checks
Just check the version of ATK (which we can reliably get via
gobject introspection). State the rest of the requirements in
a README. Anyone who is building dependencies themselves will
surely read the README; everyone else should be running a quite
recent version of their chosen distro if they expect to have
results which truly reflect the implementation being tested.